Upgrades, now has intrinsics and extrinsics

+++ tools/nbs_tools/camera_calibration/model/equidistant.py
@@ -4,11 +4,11 @@ import tensorflow as tf
 
 
 class Equidistant(tf.keras.Model):
-    def __init__(self, focal_length, centre, **kwargs):
-        super(Equidistant, self).__init__()
+    def __init__(self, focal_length, centre, dtype=tf.float64, **kwargs):
+        super(Equidistant, self).__init__(dtype=dtype)
 
-        self.focal_length = self.add_weight("focal_length", shape=(), initializer="ones", dtype=tf.float32)
-        self.centre = self.add_weight("centre", shape=(2), initializer="zeros", dtype=tf.float32)
+        self.focal_length = self.add_weight("focal_length", shape=(), initializer="ones", dtype=self.dtype)
+        self.centre = self.add_weight("centre", shape=(2), initializer="zeros", dtype=self.dtype)
 
         self.focal_length.assign(focal_length)
         self.centre.assign(centre)
